+++ app/example_app/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 example_app = Blueprint('example_app', __name__)
 
-from . import views, errors
+from . import views, errors, announcement_board
