+++ app/templates/announcement_board/get_all_announcements.html
@@ -1,4 +1,4 @@
-{% extends "announcement_board/layout.html" %}
+{% extends "layout.html" %}
 {% block title %}All anouncements{% endblock %}
 {% block head %}
     {{ super() }}
@@ -19,6 +19,7 @@
                 <td>{{ announcement.title }}
                 <td>{{ announcement.description }}
                 <td>{{ announcement.date.strftime('%Y-%m-%d') }}
+                <td><a href={{ url_for('example_app.get_announcement', id=announcement.id) }}>Szczegóły ogłoszenia {{ announcement.id }}</a>
       {%- endfor %}
     </table>
 {% endblock %}
