+++ manage.py
@@ -1,6 +1,6 @@
 import os
 from app import create_app, db
-from app.models import User, Role
+from app.models import User, Role, Announcement
 from flask_script import Manager, Shell
 from flask_migrate import Migrate, MigrateCommand
 
@@ -9,7 +9,7 @@ manager = Manager(app)
 migrate = Migrate(app, db)
 
 def make_shell_context():
-    return dict(app=app, db=db, User=User, Role=Role)
+    return dict(app=app, db=db, User=User, Role=Role, Announcement=Announcement)
 
 
 manager.add_command('shell', Shell(make_context=make_shell_context))
